refactor: move `BoundBy` from `Ast` to `shared`
