+++ main/src/ca/uwaterloo/flix/language/phase/EntryPoints.scala
@@ -513,7 +513,7 @@ object EntryPoints {
       * }}}
       */
     private def mkEntryPoint(oldEntryPoint: TypedAst.Def, root: TypedAst.Root)(implicit flix: Flix): TypedAst.Def = {
-      val argSym = Symbol.freshVarSym("_unit", Ast.BoundBy.FormalParam, SourceLocation.Unknown)
+      val argSym = Symbol.freshVarSym("_unit", BoundBy.FormalParam, SourceLocation.Unknown)
 
       // The new type is `Unit -> Unit \ IO + eff` where `eff` is the effect of the old main.
       val argType = Type.Unit
